BACKLOG-10165 remove fragments and graphql dependency from NodeTree
* Keep the Picker component configurable with fragments
