+++ packages/react-apollo/src/picker/Picker.jsx
@@ -5,11 +5,44 @@ import * as _ from 'lodash';
 import {PredefinedFragments, replaceFragmentsInDocument} from '@jahia/apollo-dx';
 import PropTypes from 'prop-types';
 
+const PickerItemsFragment = {
+    mixinTypes: {
+        applyFor: 'node',
+        variables: {
+            lang: 'String!'
+        },
+        gql: gql`fragment MixinTypes on JCRNode {
+            mixinTypes {
+                name
+            }
+        }`
+    },
+    isPublished: {
+        applyFor: 'node',
+        variables: {
+            lang: 'String!'
+        },
+        gql: gql`fragment PublicationStatus on JCRNode {
+            publicationStatus: aggregatedPublicationInfo(language: $lang) {
+                publicationStatus
+            }
+        }`
+    },
+    primaryNodeType: {
+        applyFor: 'node',
+        gql: gql`fragment PrimaryNodeTypeName on JCRNode {
+            primaryNodeType {
+                name
+            }
+        }`
+    }
+};
+
 class Picker extends React.Component {
     constructor(props) {
         super(props);
 
-        const {
+        let {
             fragments,
             rootPaths,
             onOpenItem,
@@ -20,6 +53,10 @@ class Picker extends React.Component {
             onSelectionChange,
             defaultOpenPaths
         } = props;
+
+        if (!fragments) {
+            fragments = [PickerItemsFragment.mixinTypes, PickerItemsFragment.primaryNodeType, PickerItemsFragment.isPublished, PredefinedFragments.displayName];
+        }
 
         this.query = gql`
             query PickerQuery($rootPaths:[String!]!, $selectable:[String]!, $openable:[String]!, $openPaths:[String!]!, $types:[String]!) {
@@ -271,8 +308,7 @@ class Picker extends React.Component {
 
         return (
             <Query query={this.query} variables={vars} fetchPolicy="cache-first">
-                {
-                ({error, loading, data, refetch}) => {
+                {({error, loading, data, refetch}) => {
                     if (setRefetch) {
                         setRefetch({
                             query: this.query,
@@ -291,7 +327,7 @@ class Picker extends React.Component {
                             return renderProp({pickerEntries: this.previousEntries, loading, ...this.eventsHandlers});
                         }
 
-                            return renderProp({pickerEntries: [], loading, ...this.eventsHandlers});
+                        return renderProp({pickerEntries: [], loading, ...this.eventsHandlers});
                     }
 
                     if (error) {
@@ -302,8 +338,7 @@ class Picker extends React.Component {
                     this.previousEntries = pickerEntries;
 
                     return renderProp({pickerEntries, loading, ...this.eventsHandlers});
-                }
-            }
+                }}
             </Query>
         );
     }
